+++ src/molecules/CurrentElectricityDelivered/CurrentElectricityDelivered.js
@@ -2,6 +2,7 @@ import React from 'react';
 import styles from './CurrentElectricityDelivered.module.css';
 
 import Skeleton from '../../atoms/Skeleton/Skeleton';
+import Message from '../../atoms/Message/Message';
 
 import { useQuery, gql } from '@apollo/client';
 
@@ -19,7 +20,7 @@ export default function CurrentElectricityDelivered() {
 		fetchPolicy: 'cache-and-network',
 	});
 
-	if (error) return <p>Error :(</p>;
+	if (error) return <Message type="error">{error.message}</Message>;
 	return (
 		<div className={styles.container}>
 			{!loading ? data.currentElectricityUsage ? data.currentElectricityUsage.delivered + ' W' : 0 + ' W' : <Skeleton width="3em" /> }