+++ src/molecules/ElectricityDeliveredChart/ElectricityDeliveredChart.js
@@ -1,5 +1,7 @@
 import React from 'react';
 
+import Message from '../../atoms/Message/Message';
+
 import { useQuery, gql } from '@apollo/client';
 
 import Highcharts from 'highcharts';
@@ -27,7 +29,7 @@ export default function ElectricityDeliveredChart({ resolution, hoursInThePast }
 	});
 
 	if (loading) return <p>Loading...</p>;
-	if (error) return <p>Error :(</p>;
+	if (error) return <Message type="error">{error.message}</Message>;
 
 	return (
 		<div>