+++ src/molecules/ElectricityReceivedWeekChart/fragments.js
@@ -1,0 +1,12 @@
+import { gql } from '@apollo/client';
+
+export const CONSUMPTION = gql`
+	fragment consumption on ElectricityConsumption {
+		received
+		delivered
+		period {
+			start
+			end
+		}
+	}
+`;