+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2021-04-18
+### Added
+- Current week Electricity received widget.
+
+### Changed
+- Better error messages for most components inside widgets.
+
+### Fixed
+- Fatal for 2 widgets when getting invalid data from api.
+
 ## [1.4.0] - 2021-04-16
 ### Added
 - Add verified water reading form, for correcting the water sensor.